fix: dropdown ref 리팩토링 hook 추가 #108
